// resources/js/types/plotAdd.ts
export interface Owner {
    image: string;
    name: string;
    dob: string;
    religion: string;
    nationality: string;
    profession: string;
    tin: string;
    email: string;
    permanentAddress: string;
    presentAddress: string;
    phoneNumber: string;
}

export interface Plot{
    id: string;
    price: number;
}